PO-2320 Fix central fund route mapping
Move central fund handling into MajorCreditorApiController and remove the duplicate controller mapping.
